DronePortGCS-integration_drone/docker/log-consumer/requirements.txt
kafka-python>=2.0.2
paho-mqtt>=2.0.0
elasticsearch>=8.0.0,<9.0.0
